Ensure the last processed block number is stored
First, the signal handler was not set up correctly, there was SIGKILL
instead of SIGTERM in the catcher, so the process was simply dying on
SIGTERM, no good. This is now fixed.

And anyway, to be sure the last processed block number is safe, the
block processor is flushing it into the database every minute now, so
even if the process crashes, the lost is not so horrible.

This fixes #5