BlockProcessor: Initialise config flusher properly
Fix a bug that was causing the BlockProcessor to sometimes start with
block 0 and start sending out old notifications. This was because the
block number was actually not being initialised properly so in case
there was an error right at the beginning, block number 1 was stored as
the next block number to be processed.