web: Add a new section: Event Stream
It is now possible to watch events online in real-time.

The implementation is for now pretty simple. You only see what is emitted when
you are connected to the WebSocket. There is no buffering, no history yet.

The event stream is implemented as a special Notifier for the notifications
subsystem. The WebSocket connection is found based on the event user ID.

Resolves #23