Add a new notifier: Steemit Chat
This notifier is basically a clone of the Slack notifier since Rocket Chat is
trying to be Slack-compatible. So the payload sent to the chat is the same.

What is obviously not the same is the way users connect with Steemit Chat. They
have to provide their credentials to authenticate and their username is stored.
SteemWatch then uses an internal account to send the messages because it is not
possible to send messages to yourself.

Resolves #30