Add new login provider: Reddit
It is now possible to sign into SteemWatch using Reddit.

Unexpected amount of work was necessary to implement this actually.

The OAuth2 package could not be used because Reddit is limiting HTTP requests by
User-Agent header and it was not possile to set custom header for that library.

Secondly, Reddit does not give us any user email, so the whole auth framework
had to be modified to allow for custom social links. But in case email address
is needed for anything in the future, Reddit users will be missing it.

Resolves #33