Remove support for descendant.published
The facility turned out to be not used very often. Considering the
management overhead, it is not worth it keeping it for now.

A more clever approach is needed, but it is yet to be uncovered. It is
too much to ask the user to insert a link for every article they
publish. The right solution to watching articles is to do something
clever by default. That is not happening right now, but it will come.